+++ frontend/src/App.tsx
@@ -1,11 +1,12 @@
-// import React from 'react';
 import { BrowserRouter as Router, Route, Routes, Navigate } from 'react-router-dom';
-import './styles/App.css';
+import { ToastContainer } from 'react-toastify';
+
 import LoginPage from './pages/LoginPage.tsx';
 import RegisterPage from './pages/RegisterPage.tsx';
 import VerifyPage from './pages/VerifyPage.tsx';
 import Dashboard from './pages/Dashboard.tsx';
-import { ToastContainer } from 'react-toastify';
+
+import './styles/App.css';
 import 'react-toastify/dist/ReactToastify.css';
 
 function App() {
@@ -20,7 +21,6 @@ function App() {
           <Route path="*" element={<Navigate to="/" />} />
         </Routes>
 
-        {/* ✅ Toasts will show up globally */}
         <ToastContainer position="top-center" autoClose={3000} aria-label="Notification Toasts" />
       </>
     </Router>
