+++ frontend/src/types/APITypes.ts
@@ -1,6 +1,8 @@
+import type { FeedPost, FriendData, LeaderboardEntry, ProfileData } from "./dashboardTypes";
+
 export interface LoginInfo {
-    accessToken: string;
-    userId: string;
+  accessToken: string;
+  userId: string;
 }
 
 export interface UserPayload {
@@ -8,4 +10,33 @@ export interface UserPayload {
   firstName: string;
   lastName: string;
   iat: number;
+}
+export interface FeedResponse {
+  feed: FeedPost[];
+  jwtToken: string;
+}
+export interface ProfileResponse {
+  profileData: ProfileData;
+  jwtToken: string;
+}
+export interface FriendsResponse {
+  friends: FriendData[];
+  jwtToken: string;
+}
+
+export interface CurrentQuestResponse {
+  success: boolean;
+  currentQuest: {
+    questId: string;
+    timestamp: string;
+    questData: Record<string, unknown>;
+    questDescription: string;
+
+  };
+  timestamp: string;
+}
+
+export interface LeaderboardResponse {
+  scoreboard: LeaderboardEntry[];
+  jwtToken: string;
 }
